Some further restructuring simplification
